+++ desktop/cmp/panel/PanelModel.js
@@ -9,6 +9,7 @@ import {XH, HoistModel} from '@xh/hoist/core';
 import {observable, action} from '@xh/hoist/mobx';
 import {withDefault} from '@xh/hoist/utils/js';
 import {start} from '@xh/hoist/promise';
+import {RenderMode} from '@xh/hoist/enums';
 import {isNil} from 'lodash';
 
 /**
@@ -58,8 +59,7 @@ export class PanelModel {
      * @param {boolean} [config.defaultCollapsed] - Default collapsed state.
      * @param {string} config.side - Side towards which the panel collapses or shrinks. This relates
      *      to the position within a parent vbox or hbox in which the panel should be placed.
-     * @param {string} [config.collapsedRenderMode] - How should collapsed content be rendered?
-     *      Valid values include 'lazy', 'always', and 'unmountOnHide'.
+     * @param {RenderMode} [config.collapsedRenderMode] - How should collapsed content be rendered?
      * @param {?string} [config.prefName] - preference name to store sizing and collapsed state.
      * @param {boolean} [config.showSplitter] - Should a splitter be rendered at the panel edge?
      * @param {boolean} [config.showSplitterCollapseButton] - Should the collapse button be visible
@@ -76,7 +76,7 @@ export class PanelModel {
         maxSize = null,
         defaultCollapsed = false,
         side,
-        collapsedRenderMode = 'lazy',
+        collapsedRenderMode = RenderMode.LAZY,
         prefName = null,
         showSplitter = resizable || collapsible,
         showSplitterCollapseButton = showSplitter && collapsible,